requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
asn1crypto==0.22.0        # via cryptography
babel==2.5.1
babeldjango==0.2.2
boto3==1.4.7
botocore==1.7.14          # via boto3, s3transfer
braintree==3.37.2         # via django-payments
certifi==2017.7.27.1      # via requests
cffi==1.10.0              # via cryptography
chardet==3.0.4            # via requests
cryptography==2.0.3       # via django-payments
defusedxml==0.5.0         # via python3-openid, social-auth-core
dj-database-url==0.4.2
dj-email-url==0.0.10
django-bootstrap3==9.0.0
django-cache-url==1.3.1
django-countries==4.6.1
django-filter==1.0.4
django-map-widgets==0.1.9
django-mptt==0.8.7
django-payments==0.11.0.3
django-prices-openexchangerates==0.1.15
django-prices==0.6.2
django-redis==4.8.0
django-render-block==0.5  # via django-templated-email
django-storages==1.6.5
django-templated-email==2.2.0
django-versatileimagefield==1.7.1
django-webpack-loader==0.5.0
django==1.11.5
dnspython==1.15.0         # via email-validator
docutils==0.14            # via botocore
elasticsearch==5.4.0
email-validator==1.0.2    # via faker
faker==0.8.3
google-i18n-address==2.0.2
google-measurement-protocol==0.1.6
graphene-django==1.2.0
graphene==1.4.1           # via graphene-django
graphql-core==1.1         # via graphene, graphql-relay
graphql-relay==0.4.5      # via graphene
idna==2.6                 # via cryptography, email-validator, requests
iso8601==0.1.12           # via graphene-django
jmespath==0.9.3           # via boto3, botocore
jsonfield==2.0.2
markdown==2.6.9
maxminddb-geolite2==2017.803
maxminddb==1.3.0
mock==2.0.0
oauthlib==2.0.4           # via requests-oauthlib, social-auth-core
olefile==0.44             # via pillow
pbr==3.1.1                # via mock
pillow==4.0.0             # via django-versatileimagefield
prices==0.5.9
promise==2.0.2            # via graphene, graphql-core, graphql-relay
psycopg2==2.7.3.1
purl==1.3.1
py==1.4.34                # via pytest
pycparser==2.18           # via cffi
pyjwt==1.5.3              # via django-payments, social-auth-core
pytest-django==3.1.2
pytest==3.2.2
python-dateutil==2.6.1    # via botocore, faker
python3-openid==3.1.0     # via social-auth-core
pytz==2017.2              # via babel, django
redis==2.10.6             # via django-redis
requests-oauthlib==0.8.0  # via social-auth-core
requests==2.18.4
s3transfer==0.1.10        # via boto3
satchless==1.1.3
singledispatch==3.4.0.3   # via graphene-django
six==1.11.0               # via cryptography, django-templated-email, faker, graphene, graphene-django, graphql-core, graphql-relay, mock, promise, purl, python-dateutil, singledispatch, social-auth-app-django, social-auth-core
social-auth-app-django==1.2.0
social-auth-core==1.4.0   # via social-auth-app-django
stripe==1.65.0            # via django-payments
suds-jurko==0.6           # via django-payments
text-unidecode==1.0
typing==3.6.2             # via promise
urllib3==1.22             # via elasticsearch, requests
uwsgi==2.0.15
xmltodict==0.11.0         # via django-payments
